refactor(tests): split ask tests to separate file for FR-006 ratchet
Move repo_map ask dispatch tests from repo-map.test.ts (already at cap)
to repo-map-ask.test.ts to satisfy the test-file line-cap ratchet.
