+++ tests/unit/tools/repo-map.test.ts
@@ -454,60 +454,6 @@ describe('repo_map: validation', () => {
 	});
 });
 
-describe('repo_map: ask', () => {
-	it('requires question', async () => {
-		await call({ action: 'build' });
-		const out = await call({ action: 'ask' });
-		const r = JSON.parse(out) as { success: boolean; error: string };
-		expect(r.success).toBe(false);
-		expect(r.error).toContain('requires `question`');
-	});
-
-	it('rejects control-char question', async () => {
-		await call({ action: 'build' });
-		const out = await call({ action: 'ask', question: 'hello\x00world' });
-		const r = JSON.parse(out) as { success: boolean; error: string };
-		expect(r.success).toBe(false);
-		expect(r.error).toContain('control characters');
-	});
-
-	it('returns hits after build', async () => {
-		await call({ action: 'build' });
-		const out = await call({ action: 'ask', question: 'add util' });
-		const r = JSON.parse(out) as {
-			success: boolean;
-			hits: Array<{ file: string; score: number }>;
-			budget: { requested: number; returned: number; dropped: number };
-		};
-		expect(r.success).toBe(true);
-		expect(r.hits.length).toBeGreaterThan(0);
-		expect(r.budget.returned).toBeGreaterThan(0);
-	});
-
-	it('rejects whitespace-only question', async () => {
-		await call({ action: 'build' });
-		const out = await call({ action: 'ask', question: '   \t  ' });
-		const r = JSON.parse(out) as { success: boolean; error: string };
-		expect(r.success).toBe(false);
-		expect(r.error).toContain('question is empty');
-	});
-
-	it('rejects question exceeding max length', async () => {
-		await call({ action: 'build' });
-		const out = await call({ action: 'ask', question: 'x'.repeat(501) });
-		const r = JSON.parse(out) as { success: boolean; error: string };
-		expect(r.success).toBe(false);
-		expect(r.error).toContain('exceeds maximum length');
-	});
-
-	it('accepts question at exactly the max length boundary', async () => {
-		await call({ action: 'build' });
-		const out = await call({ action: 'ask', question: 'x'.repeat(500) });
-		const r = JSON.parse(out) as { success: boolean };
-		expect(r.success).toBe(true);
-	});
-});
-
 describe('repo_map: callers / dead_exports', () => {
 	it('callers reports files that reference an exported symbol', async () => {
 		await call({ action: 'build' });
